Move XHR service into a separate implementation file.
Another one of #1055
