+++ test/size.txt
@@ -1,9 +1,9 @@
       max   |         min   |       gzip   |   file
       ---   |         ---   |        ---   |   ---
-616.01 kB   |   157.78 kB   |   43.76 kB   |   v0.js / amp.js
- 330.7 kB   |    41.21 kB   |   13.26 kB   |   v0/amp-access-0.1.js
+629.49 kB   |   161.32 kB   |   44.95 kB   |   v0.js / amp.js
+ 318.3 kB   |    37.76 kB   |   12.15 kB   |   v0/amp-access-0.1.js
  27.06 kB   |     5.96 kB   |    2.35 kB   |   v0/amp-accordion-0.1.js
-358.51 kB   |    58.33 kB   |   20.16 kB   |   v0/amp-analytics-0.1.js
+346.11 kB   |    54.63 kB   |   19.06 kB   |   v0/amp-analytics-0.1.js
  87.73 kB   |     9.49 kB   |    3.62 kB   |   v0/amp-anim-0.1.js
  84.57 kB   |     7.22 kB   |    2.92 kB   |   v0/amp-audio-0.1.js
  79.42 kB   |     8.33 kB   |    3.16 kB   |   v0/amp-brid-player-0.1.js
@@ -21,9 +21,9 @@
  78.85 kB   |     7.79 kB   |    3.03 kB   |   v0/amp-jwplayer-0.1.js
  92.84 kB   |     8.66 kB   |    3.37 kB   |   v0/amp-kaltura-player-0.1.js
 137.26 kB   |    15.35 kB   |    4.83 kB   |   v0/amp-lightbox-0.1.js
-114.42 kB   |    13.52 kB   |    4.92 kB   |   v0/amp-list-0.1.js
+100.73 kB   |     9.04 kB   |     3.4 kB   |   v0/amp-list-0.1.js
 143.99 kB   |     40.5 kB   |   14.27 kB   |   v0/amp-mustache-0.1.js
-120.91 kB   |    24.95 kB   |     7.6 kB   |   v0/amp-pinterest-0.1.js
+107.21 kB   |    20.46 kB   |    6.04 kB   |   v0/amp-pinterest-0.1.js
  71.11 kB   |     6.52 kB   |    2.61 kB   |   v0/amp-reach-player-0.1.js
  95.78 kB   |     11.6 kB   |    4.16 kB   |   v0/amp-sidebar-0.1.js
 106.34 kB   |    12.75 kB   |    4.54 kB   |   v0/amp-slides-0.1.js
@@ -31,7 +31,7 @@
  71.69 kB   |     6.67 kB   |    2.66 kB   |   v0/amp-soundcloud-0.1.js
  79.66 kB   |     8.51 kB   |    3.14 kB   |   v0/amp-springboard-player-0.1.js
 133.14 kB   |    14.57 kB   |    5.74 kB   |   v0/amp-twitter-0.1.js
-251.19 kB   |    17.86 kB   |     6.2 kB   |   v0/amp-user-notification-0.1.js
+237.49 kB   |    13.43 kB   |    4.73 kB   |   v0/amp-user-notification-0.1.js
  71.62 kB   |     6.62 kB   |    2.65 kB   |   v0/amp-vimeo-0.1.js
  71.15 kB   |     6.49 kB   |    2.61 kB   |   v0/amp-vine-0.1.js
  82.11 kB   |      8.8 kB   |    3.38 kB   |   v0/amp-youtube-0.1.js
